rootdir: include /etc/recovery.fstab in recovery
A few commits before had moved /etc/recovery.fstab to /etc/fstab,
saying that missing /etc/fstab makes /cache not mounted automatically
in case there isn't any update on /cache.

However, missing /etc/recovery.fstab cause the recovery's updater to be
unable to find the update reside in /cache on boot. Beside, this file
is needed by system-image-upgrader to determine what partitions are
available on the system. Missing this file cause system-image-upgrader
to be unable to flash raw partitions.

The solution is to include both files in recovery image. As both files
are actually the same content, I simply tell the build system to copy
the same file twice.

Change-Id: Id780fafd072db67e6e5b8d0debd2bc2681772d69